fixed my get decimal value.

+++ week03/Fractions/Fractions.cs
@@ -29,7 +29,7 @@ class Fractions
 
     public double GetDecimalValue()
     {
-        double value = _top/_bottom;
+        double value = (double)_top / (double)_bottom;
         return value;
     }
 }
